+++ packages/utils/src/index.ts
@@ -72,6 +72,7 @@ import rangeSpec from './rangeSpec.ts';
 import removeOptionalEmptyObjects from './removeOptionalEmptyObjects.ts';
 import replaceStringParameters from './replaceStringParameters.ts';
 import resolveUiSchema from './resolveUiSchema.ts';
+import schemaHasNestedConditional from './schemaHasNestedConditional.ts';
 import schemaRequiresTrueValue from './schemaRequiresTrueValue.ts';
 import SelectedOptionDescription from './SelectedOptionDescription.tsx';
 import type { SelectedOptionDescriptionProps } from './SelectedOptionDescription.tsx';
@@ -187,6 +188,7 @@ export {
   removeOptionalEmptyObjects,
   replaceStringParameters,
   resolveUiSchema,
+  schemaHasNestedConditional,
   schemaRequiresTrueValue,
   setByPath,
   SelectedOptionDescription,